+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Feature :
   - Timer duration is now saved for each timer type
   - Better Jinx display in the reference modal
+  - Select all button in script builder
 - Bug fixes :
   - Reminders match the color of their role's team
   - Printing the correct type of the special votes (for now only CL) in the vote history